pyproject.toml
[tool.poetry]
name = "cierto_2"
version = "0.1.0"
description = ""
authors = ["Yuriy Trofimov <70945999+yytrofimov@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.1"
Flask-Caching = "^1.10.1"
Flask-Mail = "^0.9.1"
qrcode = {extras = ["pil"], version = "^7.3"}
python-dotenv = "^0.19.0"
Flask-WTF = "^0.15.1"
SQLAlchemy = "^1.4.23"
psycopg2 = "^2.9.1"
redis = "^3.5.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
